requirements.txt
RPi.GPIO
opencv-python
requests
aiohttp
pyyaml
python-dotenv